+++ p2p/listener_test.go
@@ -9,7 +9,7 @@ import (
 
 func TestListener(t *testing.T) {
 	// Create a listener
-	l := NewDefaultListener("tcp", ":8001", true, log.TestingLogger())
+	l := NewDefaultListener("tcp", ":8001", false, log.TestingLogger())
 
 	// Dial the listener
 	lAddr := l.ExternalAddress()