+++ CHANGELOG.md
@@ -9,7 +9,7 @@ BUG FIXES:
   (`rpc.grpc_max_open_connections`). Check out [Running In
   Production](https://tendermint.readthedocs.io/en/master/running-in-production.html)
   guide if you want to increase them.
-  
+
 ## 0.21.2
 
 IMPROVEMENT
@@ -24,6 +24,7 @@ BREAKING CHANGES
 
 - [config] Change default ports from 4665X to 2665X. Ports over 32768 are
   ephemeral and reserved for use by the kernel.
+- [config] Rename `skip_upnp` to `upnp`, and change default to enable upnp
 - [cmd] `unsafe_reset_all` removes the addrbook.json
 
 IMPROVEMENT